docs (storage): Add host and port fields for postgres
Just did a recent deployment of Dex one one of the clusters and it
took me a while to find the proper configuration. My initial
impression was that the field database should contain the host and
port information along with database name in some form of URI which
isn't correct. Having the host & port field would have saved me some
time.

Signed-off-by: Sibi Prabakaran <sibi@psibi.in>